fix(build): updated to webpack 4 and Angular 5
The build and demo run, but webpack 4 requires Angular 5 due to typescript 2.7. The tests do not run, which I believe is due to lack of support for webpack 4 and awesome-typescript-loader. There is a new awesome-typescript-loader version in the works, so I’ll push this to a feature branch for now.

See https://github.com/s-panferov/awesome-typescript-loader/issues/534